python/ai-service/requirements.txt
# AI Service Dependencies
# Python 3.10+

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI Providers
langchain==0.1.0
langchain-anthropic==0.1.1
langchain-openai==0.0.2
anthropic==0.18.1
openai==1.6.1

# Ollama support
ollama==0.1.3

# AI Orchestrator - Multi-provider support
litellm==1.28.0

# Database & Caching
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-json-logger==2.0.7
tenacity==8.2.3  # Retry logic

# Agent Framework
langgraph>=0.0.20
apscheduler>=3.10.0
tzlocal>=5.0  # Required by apscheduler

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
